+++ duva/tests/cluster_ops/test_cluster_forget_when_wrong_id_given.rs
@@ -1,3 +1,5 @@
+use std::{thread::sleep, time::Duration};
+
 /// Cluster forget {node id} is used in order to remove a node, from the set of known nodes for the node receiving the command.
 /// In other words the specified node is removed from the nodes table of the node receiving the command.
 use crate::common::{Client, ServerEnv, spawn_server_process};
@@ -10,6 +12,8 @@ fn run_cluster_forget_node_return_error_when_wrong_id_given(env: ServerEnv) -> a
     let mut client_handler = Client::new(leader_p.port);
     let replica_id = "localhost:19933";
     let cmd = format!("cluster forget {}", &replica_id);
+
+    sleep(Duration::from_millis(500));
 
     // THEN
     assert_eq!(client_handler.send_and_get(&cmd), "(error) No such peer");
